requirements.txt
mysql_connector_repackaged==0.3.1
openpyxl==3.0.10
Pillow==9.2.0
pywin32==304
tk==0.1.0